requirements.txt
# Dependencias principales del sistema Arcadium Automation
# Core asincrónico y procesamiento de cadenas

# HTTP asíncrono
aiohttp>=3.9.0

# Validación y configuración
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonschema>=4.19.0

# Utilidades
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=24.1.0
psutil>=5.9.0
watchdog>=3.0.0
requests>=2.31.0
websockets>=12.0

# Persistencia (opcional)
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Monitoreo
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# LangChain & AI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
supabase>=2.3.0
mcp>=0.9.0

# Web Framework (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP cliente para APIs externas
httpx>=0.25.0

# Base de datos PostgreSQL
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# LangChain moderno (sin deprecated)
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
langchain-postgres>=0.0.9

# Google Calendar API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.1
google-auth>=2.0.0


